requirements.txt
Django==4.2
djangorestframework
django-celery
redis
django-redis
psycopg2-binary==2.9.9
